pyproject.toml
[project]
name = "trader"
version = "0.1.0"
description = "Advanced stocks and crypto trading bot system with multiple strategies"
authors = [
    {name = "nir",email = "nir12127@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alpaca-trade-api>=3.0.0",
    "ccxt>=4.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "ta>=0.11.0",
    "python-binance>=1.0.0",
    "yfinance>=0.2.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.7.0",
    "plotly>=5.14.0",
    "websocket-client>=1.6.0",
    "colorlog>=6.7.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
]

[tool.poetry]
packages = [{include = "trader", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
